backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
supabase>=2.15.0,<3.0.0
pydantic-settings>=2.8.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
email-validator>=2.2.0,<3.0.0
pypdf>=5.4.0,<6.0.0
openai>=1.72.0,<2.0.0
youtube-transcript-api>=0.6.2,<1.0.0